+++ rich/theme.py
@@ -15,7 +15,9 @@ class Theme:
 
     styles: Dict[str, Style]
 
-    def __init__(self, styles: Mapping[str, StyleType] = None, inherit: bool = True):
+    def __init__(
+        self, styles: Optional[Mapping[str, StyleType]] = None, inherit: bool = True
+    ):
         self.styles = DEFAULT_STYLES.copy() if inherit else {}
         if styles is not None:
             self.styles.update(
@@ -35,7 +37,7 @@ class Theme:
 
     @classmethod
     def from_file(
-        cls, config_file: IO[str], source: str = None, inherit: bool = True
+        cls, config_file: IO[str], source: Optional[str] = None, inherit: bool = True
     ) -> "Theme":
         """Load a theme from a text mode file.