+++ CHANGELOG.md
@@ -5,7 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [10.0.0] - Unreleased
+## [10.1.1] - Unreleased
+
+### Added
+
+- Added syntax for call, i.e. "Foo(bar)" will highlight Foo.pyth
+
+## [10.1.0] - 2020-04-03
+
+### Fixed
+
+- Fixed support for jupyter qtconsole and similar Jupyter environments
+
+## [10.0.1] - 2021-03-30
+
+### Fixed
+
+- Fixed race condition that duplicated lines in progress https://github.com/willmcgugan/rich/issues/1144
+
+## [10.0.0] - 2021-03-27
 
 ### Changed